requirements.txt
Flask
flask-cors
python-dotenv
langchain
langchain-openai
langchain-huggingface
langchain-pinecone
pinecone-client
openai
